requirements.txt
pillow
qrcode
XlsxWriter
opencv-python-headless
pandas
openpyxl